Fix incorrect package name in dependencies
This commit fixes the package name error that caused the build to fail:
```
E: Unable to locate package libpangoft2-1.0-dev
```

Changes made:
- Removed the non-existent `libpangoft2-1.0-dev` package
- Kept `libpango1.0-dev` which already includes the PangoFT2 functionality

The `libpango1.0-dev` package in Ubuntu provides all the necessary Pango libraries including PangoFT2, so we don't need a separate package for it.

This simple fix should allow the build to proceed past the system dependency installation phase.